# reference/3.1.0/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SpatialService REST services",
    "description": "REST services for interacting with the spatial-service webapp",
    "termsOfService": "",
    "contact": {
      "name": "Support",
      "email": "support@l-a.site"
    },
    "license": {
      "name": "Mozilla Public License 1.1",
      "url": "https://www.mozilla.org/en-US/MPL/1.1/"
    },
    "version": "3.1.0"
  },
  "servers": [
    {
      "url": "https://spatial.l-a.site/ws"
    }
  ],
  "paths": {
...
    "/fields": {
      "get": {
        "tags": [
          "fields"
        ],
        "summary": "Get list of fields",
        "description": "Get a list of fields",
        "operationId": "getFields",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "restrict to field names that contain this value and include layer information",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fields"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fields/search": {
      "get": {
        "tags": [
          "fields"
        ],
        "summary": "search for fields",
        "operationId": "searchFields",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "restrict to field names that contain this value",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fields"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/fields/layers/search": {
      "get": {
        "tags": [
          "fields"
        ],
        "summary": "search for layers",
        "operationId": "searchLayers",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "restrict to layer names that contain this value",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of layers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Layers"
                }
              }
            }
          }
        }
      }
    },
    "/field/{id}": {
      "get": {
        "tags": [
          "fields"
        ],
        "summary": "Get a field by Id",
        "description": "Get a field by Id. Includes all objects associated with the field.",
        "operationId": "getFieldById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the field",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "starting index for associated objects",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "number of associated objects to return",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "restrict to associated object names that contain this value",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fields"
                }
              }
            }
          }
        }
      }
    },
...
    "/intersect/{ids}/{lat}/{lng}": {
      "get": {
        "tags": [
          "intersect"
        ],
        "summary": "Get intersection of a point and one or more fields",
        "operationId": "intersectPoint",
        "parameters": [
          {
            "name": "ids",
            "in": "path",
            "description": "Comma delimited list of field IDs",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cl22"
          },
          {
            "name": "lat",
            "in": "path",
            "description": "Latitude",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": -22
          },
          {
            "name": "lng",
            "in": "path",
            "description": "Longitude",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 134
          }
        ],
        "responses": {
          "200": {
            "description": "Intersection results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "example": null
              }
            }
          }
        }
      }
    },
...
    "/intersect/reloadconfig": {
      "get": {
        "tags": [
          "cache"
        ],
        "summary": "Clear caches",
        "operationId": "clearCache",
        "responses": {
          "200": {
            "description": "Notification that the reload of config was successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "deprecated": true,
        "security": [
          {
            "openIdConnect": []
          }
        ]
      }
    },
...
    "/layers/search": {
      "get": {
        "tags": [
          "layers"
        ],
        "summary": "search for layers",
        "operationId": "searchLayers_1",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "restrict to layer names that contain this value",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of layers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Layers"
                }
              }
            }
          }
        }
      }
    },
...
    "/layer/{id}": {
      "get": {
        "tags": [
          "layers"
        ],
        "summary": "Get a layer by layer ID",
        "operationId": "getLayerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Layer ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Layer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Layers"
                }
              }
            }
          }
        }
      }
    },
...
    "/objects/{fid}": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Get a list of objects in a field",
        "operationId": "objectsInField",
        "parameters": [
          {
            "name": "fid",
            "in": "path",
            "description": "Field ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Paging start index",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Paging page size",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spatial Object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpatialObjects"
                  }
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
...
    "/object/{pid}": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Get a spatial object",
        "operationId": "getSpatialObject",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "description": "Id of the checklist item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spatial Object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpatialObjects"
                }
              }
            }
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Search all objects by name",
        "operationId": "searchObjects",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search term",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Paging page size",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Paging start index",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "include",
            "in": "query",
            "description": "Field IDs to search, comma delimited",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exclude",
            "in": "query",
            "description": "Field IDs to exclude from search, comma delimited",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "checklist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Distributions"
                }
              }
            }
          }
        }
      }
    },
    "/shape/upload/{pid}": {
      "delete": {
        "tags": [
          "uploads"
        ],
        "summary": "Delete an object",
        "operationId": "deleteObject",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "description": "Object ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Object with the area id",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "openIdConnect": []
          }
        ]
      }
    },
...
    "/shapes/kml/{pid}": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Get KML for an object",
        "operationId": "getKML",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "description": "Object ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "KML",
            "content": {
              "plain/text": {
                "schema": {
                  "$ref": "#/components/schemas/Distributions"
                }
              }
            }
          }
        }
      }
    },
...
    "/shape/upload/geojson": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Create an object from GeoJSON",
        "operationId": "uploadGeoJSON",
        "requestBody": {
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "Object with the area id",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "openIdConnect": []
          }
        ]
      }
    },
...
    "/shape/upload/wkt": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Create an object from WKT",
        "operationId": "uploadWkt_1",
        "requestBody": {
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "Object with the area id",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "openIdConnect": []
          }
        ]
      }
    },
...
    "/tasks/cancel/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancel a task",
        "operationId": "cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task status",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "openIdConnect": []
          }
        ]
      }
    },
...
    "/tasks/create": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create a task with JSON body or query parameters",
        "operationId": "create",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Task name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inputs",
            "in": "query",
            "description": "task inputs as JSON",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task status",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "openIdConnect": []
          }
        ]
      }
    },
    "/tasks/status/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Task status",
        "operationId": "status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task status",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Fields": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "spid": {
            "type": "string"
          },
          "sname": {
            "type": "string"
          },
          "sdesc": {
            "type": "string"
          },
          "indb": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "last_update": {
            "type": "string",
            "format": "date-time"
          },
          "namesearch": {
            "type": "boolean"
          },
          "defaultlayer": {
            "type": "boolean"
          },
          "intersect": {
            "type": "boolean"
          },
          "layerbranch": {
            "type": "boolean"
          },
          "analysis": {
            "type": "boolean"
          },
          "addtomap": {
            "type": "boolean"
          },
          "requestedId": {
            "type": "string"
          },
          "number_of_objects": {
            "type": "integer",
            "format": "int32"
          },
          "layer": {
            "$ref": "#/components/schemas/Layers"
          },
          "wms": {
            "type": "string"
          },
          "objects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpatialObjects"
            }
          },
          "layerId": {
            "type": "object"
          }
        }
      },
      "Layers": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "uid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "displayname": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "displaypath": {
            "type": "string"
          },
          "scale": {
            "type": "string"
          },
          "extents": {
            "type": "string"
          },
          "minlatitude": {
            "type": "number",
            "format": "double"
          },
          "minlongitude": {
            "type": "number",
            "format": "double"
          },
          "maxlatitude": {
            "type": "number",
            "format": "double"
          },
          "maxlongitude": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "environmentalvaluemin": {
            "type": "string"
          },
          "environmentalvaluemax": {
            "type": "string"
          },
          "environmentalvalueunits": {
            "type": "string"
          },
          "lookuptablepath": {
            "type": "string"
          },
          "metadatapath": {
            "type": "string"
          },
          "classification1": {
            "type": "string"
          },
          "classification2": {
            "type": "string"
          },
          "mddatest": {
            "type": "string"
          },
          "citation_date": {
            "type": "string"
          },
          "datalang": {
            "type": "string"
          },
          "mdhrlv": {
            "type": "string"
          },
          "respparty_role": {
            "type": "string"
          },
          "licence_level": {
            "type": "string"
          },
          "licence_link": {
            "type": "string"
          },
          "licence_notes": {
            "type": "string"
          },
          "source_link": {
            "type": "string"
          },
          "keywords": {
            "type": "string"
          },
          "path_orig": {
            "type": "string"
          },
          "path_1km": {
            "type": "string"
          },
          "path_250m": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "pid": {
            "type": "string"
          },
          "dt_added": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Fields"
            }
          },
          "requestedId": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "openIdConnect": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://auth-test.ala.org.au/cas/oidc/.well-known"
      },
      "oauth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://auth.l-a.site/cas/oidc/token",
            "refreshUrl": "https://auth.l-a.site/cas/oidc/refresh",
            "scopes": {}
          },
          "authorizationCode": {
            "authorizationUrl": "https://auth.l-a.site/cas/oidc/authorize",
            "tokenUrl": "https://auth.l-a.site/cas/oidc/token",
            "refreshUrl": "https://auth.l-a.site/cas/oidc/refresh",
            "scopes": {}
          }
        }
      }
    }
  }
}